+++ cmd/hidden_lake/_template/internal/handler/index.go
@@ -1,0 +1,20 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/number571/go-peer/internal/api"
+	http_logger "github.com/number571/go-peer/internal/logger/http"
+	"github.com/number571/go-peer/pkg/logger"
+
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/_template/pkg/settings"
+)
+
+func HandleIndexAPI(pLogger logger.ILogger) http.HandlerFunc {
+	return func(pW http.ResponseWriter, pR *http.Request) {
+		logBuilder := http_logger.NewLogBuilder(hl_t_settings.CServiceName, pR)
+		pLogger.PushInfo(logBuilder.WithMessage(http_logger.CLogSuccess))
+
+		api.Response(pW, http.StatusOK, hl_t_settings.CTitlePattern)
+	}
+}