+++ cmd/hidden_lake/_template/internal/handler/index_test.go
@@ -1,0 +1,37 @@
+package handler
+
+import (
+	"net/http"
+	"testing"
+	"time"
+
+	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/_template/pkg/client"
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/_template/pkg/settings"
+	testutils "github.com/number571/go-peer/test/_data"
+)
+
+func TestHandleIndexAPI(t *testing.T) {
+	t.Parallel()
+
+	service := testRunService(testutils.TgAddrs[47])
+	defer service.Close()
+
+	time.Sleep(100 * time.Millisecond)
+	hleClient := hl_t_client.NewClient(
+		hl_t_client.NewRequester(
+			"http://"+testutils.TgAddrs[47],
+			&http.Client{Timeout: time.Second / 2},
+		),
+	)
+
+	title, err := hleClient.GetIndex()
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	if title != hl_t_settings.CTitlePattern {
+		t.Error("incorrect title pattern")
+		return
+	}
+}