+++ cmd/hidden_lake/_template/pkg/client/client.go
@@ -1,0 +1,27 @@
+package client
+
+import (
+	"fmt"
+)
+
+var (
+	_ IClient = &sClient{}
+)
+
+type sClient struct {
+	fRequester IRequester
+}
+
+func NewClient(pRequester IRequester) IClient {
+	return &sClient{
+		fRequester: pRequester,
+	}
+}
+
+func (p *sClient) GetIndex() (string, error) {
+	res, err := p.fRequester.GetIndex()
+	if err != nil {
+		return "", fmt.Errorf("get index (client): %w", err)
+	}
+	return res, nil
+}