+++ cmd/hidden_lake/_template/pkg/client/types.go
@@ -1,0 +1,11 @@
+package client
+
+type IClient interface {
+	GetIndex() (string, error)
+	// TODO: need implementation
+}
+
+type IRequester interface {
+	GetIndex() (string, error)
+	// TODO: need implementation
+}