+++ cmd/hidden_lake/_template/pkg/settings/settings.go
@@ -1,0 +1,18 @@
+package settings
+
+const (
+	CServiceName  = "HL_T"
+	CTitlePattern = "go-peer/hidden-lake-_template"
+)
+
+const (
+	CPathYML = "hl_t.yml"
+)
+
+const (
+	CDefaultHTTPAddress = "127.0.0.1:9501"
+)
+
+const (
+	CHandleIndexPath = "/api/index"
+)